+++ tests/content/entity/light/spot/create/test.js
@@ -1,4 +1,7 @@
-var autoTester = Script.require("https://raw.githubusercontent.com/highfidelity/hifi_tests/master/tests/utils/autoTester.js" );
+var user = "highfidelity/";
+var repository = "hifi_tests/";
+var branch = "master/";
+var autoTester = Script.require("https://raw.githubusercontent.com/" + user + repository + branch + "tests/utils/autoTester.js");
 
 autoTester.perform("Spot light", Script.resolvePath("."), function(testType) {
     var spectatorCameraConfig = autoTester.setupTest();
