+++ tests/content/entity/material/create/test.js
@@ -1,4 +1,7 @@
-var autoTester = Script.require("../../../../utils/autoTester.js?raw=true");
+var user = "highfidelity/";
+var repository = "hifi_tests/";
+var branch = "master/";
+var autoTester = Script.require("https://raw.githubusercontent.com/" + user + repository + branch + "tests/utils/autoTester.js");
 
 autoTester.enableAuto(10000);
